src/main/webapp/app/entities/qr-code/delete/qr-code-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { QrCodeService } from '../service/qr-code.service';

import { QrCodeDeleteDialogComponent } from './qr-code-delete-dialog.component';

describe('QrCode Management Delete Component', () => {
  let comp: QrCodeDeleteDialogComponent;
  let fixture: ComponentFixture<QrCodeDeleteDialogComponent>;
  let service: QrCodeService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, QrCodeDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(QrCodeDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(QrCodeDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(QrCodeService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
